src/plugins/generate-ad-images/skills/generate-ad-images/scripts/requirements.txt
# Python dependencies for generate-ad-images skill

# Google Generative AI SDK for Gemini API
google-genai>=1.32.0

# PIL for image processing
Pillow>=10.0.0
